Add 'require()' to interface in print.slang (#301)
Close #6764.

Slang require to put require() capability keyword on the interface,
because we can not propagate the requirement from the implementation to
interface. So the shader needs to be changed.

We need to develop mechanism for Metal specific to input the buffer
length instead of using GetDimensions method to make this working on Metal.