Automatic testing of slangpy examples (#335)
* Add SLANGPY_DEVICE_TYPE_OVERRIDE

Allows overriding the device type when creating a slangpy device.
This is used for automatic testing of the slangpy examples.

* Initialize RHI logger with same log level as the global logger

* Update samples

* Add ability to test examples to ci.yml

* Enable testing examples in CI workflow

* set PYTHONPATH in ci.py

The old setpath scripts are not needed anymore as we install the extension directly into the source tree. However PYTHONPATH still needs to be set if the working directory does not match the project dir.

* Update samples

* Update samples