Change the _trampoline signature (#337)
* Change the _trampoline signature

Close #111

Change the _trampoline signature to exclude the CallData parameter,
because there is already a ParameterBlock created for CallData.

In addition, since we choose a very common name for the parameter name
of CallData - "data", change the context to a non-common name '__slangpy_context__'
this prevents user from writing function with the same parameter name.